Implemented transaction service

+++ Application/Interfaces/Repositories/ITransactionRepository.cs
@@ -6,7 +6,7 @@ public interface ITransactionRepository
 {
     Task<OperationResult<TransactionDto?>> AddTransactionAsync(TransactionDto transaction);
     Task<OperationResult<TransactionDto?>> UpdateTransactionAsync(TransactionDto transaction);
-    Task<OperationResult<TransactionDto?>> DeleteTransactionAsync(int id);
+    Task<OperationResult<TransactionDto?>> DeleteTransactionAsync(long id);
     Task<OperationResult<List<TransactionDto>>> GetTransactionsAsync(TransactionFilterDto filter);
     Task<OperationResult<TransactionDto?>> GetTransaction(long id);
 }
